fix: type stdin offset and reattach fromSeq as uint64
The reference declares StdinParams.Offset as *uint64 and ReattachParams.FromSeq
as uint64. claustrum used *int / int, inverting the reference at both ends of
the range: a negative value was accepted where the reference rejects it, and
2^64-1 was rejected where the reference treats it as a legal offset.

The negative case is worse than a missing validation. applyStdin computes
skip = applied - offset, so offset:-1 against applied:0 yields skip=1 and the
payload is resliced data[1:] - the first byte of the child's stdin is silently
dropped, and "applied" reports one less than was sent. offset:-N drops N bytes.

Propagate uint64 through the seq/applied lineage: streamFrame.Seq,
managedProc.seq, stdinApplied, applyStdin, reattach, and the FirstSeq/LastSeq/
StdinApplied/Applied result fields. Field ORDER is untouched and every value is
non-negative, so the marshaled frames are unchanged - the existing goldens are
byte-identical.

+++ methods_process.go
@@ -71,7 +71,7 @@ type stdinParams struct {
 	// idempotency contract (advertised as "process.stdin.offset"). A pointer so an
 	// absent field ("append here", the legacy behavior) is distinct from offset:0
 	// (which, once anything has been applied, is a duplicate). See applyStdin.
-	Offset *int `json:"offset"`
+	Offset *uint64 `json:"offset"`
 }
 
 func (s *server) processStdin(req *request) response {
@@ -157,7 +157,7 @@ func (s *server) processKillAndWait(req *request) response {
 
 type reattachParams struct {
 	ID      string `json:"id"`
-	FromSeq int    `json:"fromSeq"`
+	FromSeq uint64 `json:"fromSeq"`
 	// WantPid mirrors spawnParams.WantPid — see there. When true and the process
 	// is found, the reply carries pid + startTime.
 	WantPid bool `json:"wantPid"`
